requirements.txt

fastapi==0.120.0
fastapi_users==15.0.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
uvicorn[standard]==0.29.0
psycopg2-binary  
asyncpg  
fastapi-users[sqlalchemy] 
pytest>=7.0.0
python-dotenv

Pillow==10.0.0
aiofiles==23.2.1
python-multipart==0.0.6

asyncssh